add login

+++ src/service/mod.rs
@@ -7,7 +7,7 @@ pub mod picture;
 pub mod user;
 
 lazy_static!{
-    static ref TOKEN_CACHE: Cache<String, i64> = Cache::builder()
+    static ref TOKEN_CACHE: Cache<String, i32> = Cache::builder()
         .time_to_idle(Duration::from_secs(60 * 60 * 24 * 7)) //if the key is not accessed for 7 days, it will be removed
         .build();
 }
